refactor(dicom): async httpx client, drop requests, split router/service (F1, D3, E3, I2)
- F1: replace blocking requests.get calls inside async DICOM endpoints with
  one lazily created, shared httpx.AsyncClient (get_dicom_http_client);
  verify comes from DICOM_TLS_VERIFY (TLS TODO stated once, not per call
  site); per-call timeouts (30/60/120s) preserved; DICOMService HTTP
  methods are now async and awaited at every call site;
  requests.RequestException handling mapped to httpx.HTTPError.
- F1: replace requests_toolbelt MultipartDecoder with a local RFC 2387
  multipart/related parser (parse_multipart_related) verified
  byte-identical to the toolbelt decoder on binary DICOM payloads;
  swap the stdlib cgi.parse_header call (module removed in Python 3.13)
  for an email.message equivalent.
- F1: convert remaining backend requests users to httpx —
  scripts/synthea_to_hapi_pipeline.py, scripts/active/imaging_tools.py,
  and scripts/active/stow_dicom_to_dcm4chee.py (sync httpx; STOW body
  posted via content= for byte-fidelity).
- D3: remove requests and requests_toolbelt from backend/requirements.txt
  (no backend imports remain; httpx already pinned).
- E3: frontend apiConfig.js DICOM defaults are now empty strings instead
  of hardcoded http://arc:8080 Docker-internal URLs; consumers
  (ImagingTab, DICOMViewer) already route imaging through the backend
  /dicom proxy and only forward these values as advisory headers.
- I2: split api/dicom/dicom_service.py into service.py (DICOMService,
  config, shared client, multipart parser, validate_uid) and router.py
  (endpoints, identical URLs); dicom_service.py deleted, registration in
  api/routers/__init__.py and scripts/CLAUDE.md reference updated.
