+++ backend/requirements.txt
@@ -3,7 +3,6 @@ fastapi>=0.104.1
 uvicorn[standard]>=0.24.0
 pydantic>=2.5.0
 sqlalchemy>=2.0.23
-requests_toolbelt>=1.0
 
 # WebSocket support
 websockets==12.0
@@ -18,7 +17,6 @@ redis[hiredis]>=5.0.0
 
 # HTTP/API
 httpx==0.25.2
-requests==2.33.1
 aiofiles==23.2.1
 aiohttp==3.13.5
